[DDW-500] Code restructuring
