+++ source/renderer/app/components/wallet/send-form/WalletSendAssetsConfirmationDialog.js
@@ -9,22 +9,22 @@ import { get } from 'lodash';
 import BigNumber from 'bignumber.js';
 import { defineMessages, intlShape, FormattedHTMLMessage } from 'react-intl';
 import vjf from 'mobx-react-form/lib/validators/VJF';
-import ReactToolboxMobxForm from '../../utils/ReactToolboxMobxForm';
-import Dialog from '../widgets/Dialog';
-import DialogCloseButton from '../widgets/DialogCloseButton';
-import globalMessages from '../../i18n/global-messages';
-import LocalizableError from '../../i18n/LocalizableError';
+import ReactToolboxMobxForm from '../../../utils/ReactToolboxMobxForm';
+import Dialog from '../../widgets/Dialog';
+import DialogCloseButton from '../../widgets/DialogCloseButton';
+import globalMessages from '../../../i18n/global-messages';
+import LocalizableError from '../../../i18n/LocalizableError';
 import styles from './WalletSendAssetsConfirmationDialog.scss';
-import { FORM_VALIDATION_DEBOUNCE_WAIT } from '../../config/timingConfig';
-import { submitOnEnter } from '../../utils/form';
-import { formattedTokenWalletAmount } from '../../utils/formatters';
-import { FormattedHTMLMessageWithLink } from '../widgets/FormattedHTMLMessageWithLink';
-import HardwareWalletStatus from '../hardware-wallet/HardwareWalletStatus';
-import LoadingSpinner from '../widgets/LoadingSpinner';
-import { HwDeviceStatuses } from '../../domains/Wallet';
-import AssetToken from '../widgets/AssetToken';
-import type { HwDeviceStatus } from '../../domains/Wallet';
-import type { WalletSummaryAsset } from '../../api/assets/types';
+import { FORM_VALIDATION_DEBOUNCE_WAIT } from '../../../config/timingConfig';
+import { submitOnEnter } from '../../../utils/form';
+import { formattedTokenWalletAmount } from '../../../utils/formatters';
+import { FormattedHTMLMessageWithLink } from '../../widgets/FormattedHTMLMessageWithLink';
+import HardwareWalletStatus from '../../hardware-wallet/HardwareWalletStatus';
+import LoadingSpinner from '../../widgets/LoadingSpinner';
+import { HwDeviceStatuses } from '../../../domains/Wallet';
+import AssetToken from '../../widgets/AssetToken';
+import type { HwDeviceStatus } from '../../../domains/Wallet';
+import type { WalletSummaryAsset } from '../../../api/assets/types';
 
 const SHOW_TOTAL_AMOUNT = false;
