+++ source/renderer/app/components/wallet/send-form/WalletSendConfirmationDialog.scss
@@ -1,4 +1,4 @@
-@import '../../themes/mixins/error-message';
+@import '../../../themes/mixins/error-message';
 
 .dialog {
   font-family: var(--font-medium);
